Make tests pass

+++ test/swe/engine/MarketTest.java
@@ -19,8 +19,8 @@ public class MarketTest {
 
     @Before
     public void setUp() throws Exception {
-        aInc = new Company("A inc.", StockType.HITECH, 300);
-        bInc = new Company("B inc.", StockType.HITECH, 300);
+        aInc = new Company("A inc.", StockType.HITECH, (float) 300.0);
+        bInc = new Company("B inc.", StockType.HITECH, (float) 300.0);
 
         fooPort = new Portfolio("Bar", 200, new RandomTrader(RandomInnerTraders.BALANCED));
         barPort = new Portfolio("Foo", 400, new RandomTrader(RandomInnerTraders.BALANCED));
